requirements.txt
gradio==5.29.1
numpy==2.2.6
onnxruntime==1.22.0
Pillow==11.2.1
pinecone==6.0.2
pymongo==4.13.0
python-dotenv==1.1.0
tensorrt_cu12==10.9.0.34
torch==2.5.1
transformers==4.49.0
torch==2.5.1
